+++ src/components/resize-observer.ts
@@ -1,0 +1,2 @@
+export * from './resize-observer/resize-observer.component.js';
+export { default } from './resize-observer/resize-observer.component.js';